move hooks in component folder
